requirements.txt
pillow==10.2.0
numpy==1.26.4
raylib==5.5.0.2